feat(plugin-vue2): split vue2-jsx as standalone plugin
