+++ packages/plugin-vue2-jsx/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @rsbuild/plugin-vue2-jsx
