+++ packages/plugin-vue2-jsx/modern.config.ts
@@ -1,0 +1,3 @@
+import baseConfig from '../../scripts/modern.base.config';
+
+export default baseConfig;
