+++ packages/plugin-vue2-jsx/tests/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "baseUrl": "./",
+    "types": ["vitest/globals"]
+  },
+  "include": ["./**/*.ts"]
+}
