+++ packages/plugin-vue2/src/index.ts
@@ -1,20 +1,9 @@
 import { deepmerge } from '@rsbuild/shared/deepmerge';
 import { VueLoaderPlugin } from 'vue-loader';
-import type { RsbuildPlugin } from '@rsbuild/core';
-import type { RsbuildPluginAPI } from '@rsbuild/webpack';
-import type { SharedRsbuildConfig } from '@rsbuild/shared';
+import type { RsbuildPlugin, RsbuildPluginAPI } from '@rsbuild/core';
 import type { VueLoaderOptions } from 'vue-loader';
 
-type VueJSXPresetOptions = {
-  compositionAPI?: boolean | string;
-  functional?: boolean;
-  injectH?: boolean;
-  vModel?: boolean;
-  vOn?: boolean;
-};
-
 export type PluginVueOptions = {
-  vueJsxOptions?: VueJSXPresetOptions;
   vueLoaderOptions?: VueLoaderOptions;
 };
 
@@ -24,34 +13,7 @@ export function pluginVue2(
   return {
     name: 'plugin-vue2',
 
-    // Remove built-in react plugins.
-    // These plugins should be moved to a separate package in the next major version.
-    remove: ['plugin-react', 'plugin-antd', 'plugin-arco'],
-
     async setup(api) {
-      api.modifyRsbuildConfig((config, { mergeRsbuildConfig }) => {
-        const rsbuildConfig: SharedRsbuildConfig = {
-          output: {
-            disableSvgr: true,
-          },
-          tools: {
-            babel(_, { addPresets }) {
-              addPresets([
-                [
-                  require.resolve('@vue/babel-preset-jsx'),
-                  {
-                    injectH: true,
-                    ...options.vueJsxOptions,
-                  },
-                ],
-              ]);
-            },
-          },
-        };
-
-        return mergeRsbuildConfig(config, rsbuildConfig);
-      });
-
       api.modifyBundlerChain(async (chain, { CHAIN_ID }) => {
         chain.resolve.extensions.add('.vue');
 
@@ -72,6 +34,27 @@ export function pluginVue2(
           .use(CHAIN_ID.USE.VUE)
           .loader(require.resolve('vue-loader'))
           .options(vueLoaderOptions);
+
+        // Handle ts syntax when using Rspack
+        if (
+          api.context.bundlerType === 'rspack' &&
+          !chain.module.rules.has(CHAIN_ID.RULE.TS)
+        ) {
+          chain.module
+            .rule(CHAIN_ID.RULE.TS)
+            .type('javascript/auto')
+            .test(/\.ts$/)
+            .use(CHAIN_ID.USE.SWC)
+            .loader('builtin:swc-loader')
+            .options({
+              sourceMap: true,
+              jsc: {
+                parser: {
+                  syntax: 'typescript',
+                },
+              },
+            });
+        }
 
         chain.plugin(CHAIN_ID.PLUGIN.VUE_LOADER_PLUGIN).use(VueLoaderPlugin);
       });
