+++ packages/plugin-vue2/tests/index.test.ts
@@ -1,14 +1,11 @@
 import { expect, describe, it } from 'vitest';
 import { createStubRsbuild } from '@rsbuild/test-helper';
-import { webpackProvider } from '@rsbuild/webpack';
-import { pluginBabel } from '@rsbuild/webpack/plugin-babel';
 import { pluginVue2 } from '../src';
 
-describe('plugins/vue', () => {
+describe('plugins/vue2', () => {
   it('should add vue-loader and VueLoaderPlugin correctly', async () => {
     const rsbuild = await createStubRsbuild({
       plugins: [pluginVue2()],
-      provider: webpackProvider,
       rsbuildConfig: {},
     });
     const config = await rsbuild.unwrapConfig();
@@ -25,36 +22,6 @@ describe('plugins/vue', () => {
           },
         }),
       ],
-      provider: webpackProvider,
-      rsbuildConfig: {},
-    });
-    const config = await rsbuild.unwrapConfig();
-
-    expect(config).toMatchSnapshot();
-  });
-
-  it('should apply jsx babel plugin correctly', async () => {
-    const rsbuild = await createStubRsbuild({
-      plugins: [pluginVue2(), pluginBabel()],
-      provider: webpackProvider,
-      rsbuildConfig: {},
-    });
-    const config = await rsbuild.unwrapConfig();
-
-    expect(config).toMatchSnapshot();
-  });
-
-  it('should allow to configure jsx babel plugin options', async () => {
-    const rsbuild = await createStubRsbuild({
-      plugins: [
-        pluginVue2({
-          vueJsxOptions: {
-            injectH: false,
-          },
-        }),
-        pluginBabel(),
-      ],
-      provider: webpackProvider,
       rsbuildConfig: {},
     });
     const config = await rsbuild.unwrapConfig();
